Add crashlogs as a feature component to SDK. Add public IP check as soon as Internet is detected the first time in FeatureInternet. Minor type cast fix in DownloadService.

+++ src/com/aviq/tv/android/sdk/feature/internet/DownloadService.java
@@ -196,7 +196,7 @@ public class DownloadService extends BaseService
 					progressData.putFloat(ResultExtras.BYTES_DOWNLOADED.name(), bytesWritten);
 					progressData.putFloat(ResultExtras.BYTES_TOTAL.name(), total);
 					progressData.putDouble(ResultExtras.DOWNLOAD_RATE_MB_PER_SEC.name(), downloadRateMbPerSec);
-					progressData.putDouble(ResultExtras.TOTAL_TIME.name(), duration);
+					progressData.putLong(ResultExtras.TOTAL_TIME.name(), duration);
 					resultReceiver.send(DOWNLOAD_PROGRESS, progressData);
 					lastIterTime = System.currentTimeMillis();
 				}
